+++ src/main/java/com/example/if_else/Reposiories/LocationRepository.java
@@ -8,8 +8,8 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface LocationRepository extends JpaRepository<Location, Integer> {
+public interface LocationRepository extends JpaRepository<Location, Long> {
 
     @Override
-    Optional<Location> findById(Integer id);
+    Optional<Location> findById(Long id);
 }
