+++ src/main/java/com/example/if_else/Servises/AnimalTypeService.java
@@ -19,7 +19,7 @@ public class AnimalTypeService {
     }
 
 
-    public ResponseEntity<AnimalType> getAnimalTypeById(Integer typeId) {
+    public ResponseEntity<AnimalType> getAnimalTypeById(Long typeId) {
 
         if (typeId == null || typeId < 0) {
             return ResponseEntity.status(400).body(null);
